+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- update `admin-ui` version to `0.132.7`
+- `Set` component to `Stack`
+- `DataGrid` component to `Table`
+- `PageTitle` and `PageActions` component to `PageHeaderTitle` and `PageHeaderActions`
+- `Toggle` component to `Swtich`
+- update on `Button` component
+- remove prop key on `createSystem`
+- changed `Search` state to `value`, `onChange` and `onClear` props
+- `Tooltip` label to text
+
 ## [0.56.1] - 2022-08-08
 
 ### Changed